+++ src/core/univ3forks/SwapBasedV3.sol
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.25;
+
+address constant swapBasedV3Factory = 0x905A650133147012390c42624eEBa4D3313bec6C;
+uint8 constant swapBasedV3ForkId = 8;